# controller/userController.js
const express = require('express');
const router = express.Router();
const userService = require('../service/userService');

router.post('/register', (req, res) => {
  const { username, password, favorecido } = req.body;
  const result = userService.registerUser({ username, password, favorecido });
  if (result.error) return res.status(400).json({ error: result.error });
  res.status(201).json(result.user);
});

router.post('/login', (req, res) => {
  const { username, password } = req.body;
  const result = userService.loginUser({ username, password });
  if (result.error) return res.status(400).json({ error: result.error });
  res.json(result.user);
});

router.get('/', (req, res) => {
  const users = userService.getAllUsers();
  res.json(users);
});

module.exports = router;